python/requirements.txt
pandas
numpy
scikit-learn
mysql-connector-python
requests
beautifulsoup4
xlrd
openpyxl
statsmodels